Split PropertyDiff out of ChangesetPropertyBadge
