+++ src/components/Changeset/index.ts
@@ -2,3 +2,4 @@ export * from './ChangesetModificationWarning';
 export * from './ChangesetBadge';
 export * from './ChangesetPropertyBadge';
 export * from './ChangesetDiffContext';
+export * from './PropertyDiff';
